requirements.txt
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
google-analytics-admin==0.23.3
google-api-core==2.24.1rc1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
numpy==2.2.1
oauthlib==3.2.2
pandas==2.2.3
proto-plus==1.26.0rc1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
tzdata==2024.2
urllib3==2.3.0
